refactor: Consolidate Ad Unit chips and grid tiles.
