+++ src/pages/Shared/components/JSONViewerComponent.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import ReactJson from '@uiw/react-json-view';
 
 const JSONViewerComponent = ({
+  // comment for line break
   src = null,
   name = '',
   collapsed = 3,
   displayObjectSize = false,
   displayDataTypes = false,
   sortKeys = false,
-  // quotesOnKeys = false,
   indentWidth = 2,
   collapseStringsAfterLength = 100,
   style,
@@ -22,7 +22,6 @@ const JSONViewerComponent = ({
       displayObjectSize={displayObjectSize}
       displayDataTypes={displayDataTypes}
       objectSortKeys={sortKeys}
-      // quotesOnKeys={quotesOnKeys}
       indentWidth={indentWidth}
       stringEllipsis={collapseStringsAfterLength}
       style={{ fontSize: '12px', fontFamily: 'roboto', padding: '15px', ...style }}
